pyproject.toml
[project]
name = "urlshorter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.8.0",
    "asyncpg>=0.30.0",
    "python-multipart>=0.0.22",
    "greenlet>=3.3.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
